requirements.txt
pytest==8.4.1
requests==2.32.3
allure-pytest==2.13.5
pytest-html==3.2.0
PyYAML==6.0.2
